Refactor with new test file for Direction class
